+++ demo/component/funnel/index.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import Basic from './basic';
+import Symmetric from './symmetric';
 
 
 export default class FunnelChart extends Component {
@@ -8,6 +9,7 @@ export default class FunnelChart extends Component {
       <div className='funnel-charts'>
         <div className='funnel-chart'>
           <Basic />
+          <Symmetric />
         </div>
       </div>
     );